requirements.txt
PyPDF2
gradio
requests
python-dotenv
regex
pdfplumber
python-docx
together